% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adults}
\alias{adults}
\title{adults
This is a Phyloseq object containing example data from a 16S amplicon study
performed on adult mosquitoes from Burkina Faso.}
\format{
An object of class \code{phyloseq} of length 1.
}
\usage{
adults
}
\description{
adults
This is a Phyloseq object containing example data from a 16S amplicon study
performed on adult mosquitoes from Burkina Faso.
}
\keyword{datasets}
