% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageCurvePlotData.R
\docType{class}
\name{AvCurvePlotData-class}
\alias{AvCurvePlotData-class}
\title{Class storing the dataframe used to plot the
parametric averaged survival curves}
\description{
Class storing the dataframe used to plot the
parametric averaged survival curves
}
\section{Slots}{

\describe{
\item{\code{plotdata}}{(data frame) to be used to plot
the survival data object it has the following columns
\itemize{
  \item{t} survival time 
  \item{S} probability of survival 
  \item{Arm} which arm the data in this row is associated with
  \item{lower} lower (95%) confidence interval for S
  \item{upper} upper (95%) confidence interval for S
  \item{model} which model was used to generate the data (or "KM" for the Kaplan-Meier estimator)
}}
}}

