% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_dist_per_sample.R
\name{avg_pairwise_dist_per_sample}
\alias{avg_pairwise_dist_per_sample}
\title{Calculate Average Pairwise Distance for Clusters Across Rarefaction Thresholds}
\usage{
avg_pairwise_dist_per_sample(
  input,
  repeats = 10,
  t_min = 50,
  t_max = 250,
  t_step = 10,
  cores = 4
)
}
\arguments{
\item{input}{A phyloseq object.}

\item{repeats}{A positive integer. The number of rarefaction repeats. Default = 10.}

\item{t_min}{An integer. The minimum rarefaction threshold. Default = 50.}

\item{t_max}{An integer. The maximum rarefaction threshold. Default = 250.}

\item{t_step}{A float. The step size for thresholds. Default = 10.}

\item{cores}{An integer. The number of cores for parallel processing. Default = 4.}
}
\value{
A single ggplot object with facets showing how the average pairwise distances change across thresholds for each sample_id.
}
\description{
Calculate Average Pairwise Distance for Clusters Across Rarefaction Thresholds
}
\examples{
avg_pairwise_dist_per_sample(HLCYG_physeq_data, repeats = 10, t_min = 50, t_max = 250, t_step = 10)
}
