% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pairwise_dist_plot.R
\name{avg_pairwise_dist_plot}
\alias{avg_pairwise_dist_plot}
\title{Plot Average Pairwise Distance for Sample Clusters Across Rarefaction Thresholds}
\usage{
avg_pairwise_dist_plot(avg_distances)
}
\arguments{
\item{input}{A dataframe of average distances obtained from \code{threshold_testing}.}
}
\value{
A ggplot2 object showing average pairwise distances across thresholds.
}
\description{
This function draws a plot of \code{avg_distances} generated from the
\code{threshold_testing} function.
}
