% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmst.R
\docType{methods}
\name{calcModelRmst}
\alias{calcModelRmst}
\alias{calcModelRmst,SurvivalModel-method}
\title{Method to calculate model restricted means}
\usage{
calcModelRmst(object, ...)

\S4method{calcModelRmst}{SurvivalModel}(object, model, times,
  class = c("data.frame", "FlexTable")[2], digits = 3, ...)
}
\arguments{
\item{object}{(SurvivalModel) A survival model - note there cannot be
any covariates and armAsFactor must be FALSE}

\item{...}{additional arguments for this generic}

\item{model}{(character) The name of the model for which to calculate the restricted mean}

\item{times}{(nuermic vector) times to calculate the restricted mean}

\item{class}{('data.frame' or "FlexTable' (default)) type of output required}

\item{digits}{(numeric default=3) if outputting a FlexTable then the number of digits
to round the entries to}
}
\value{
(data.frame or FlexTable)
}
\description{
Method to calculate model restricted means
}

