% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\docType{methods}
\name{calcRmst}
\alias{calcRmst}
\alias{calcRmst}
\alias{calcRmst,SemiParametricModel-method}
\title{Method to calculate RMST on subset of data contained in a
SemiParametricModel object}
\usage{
calcRmst(object, ...)

\S4method{calcRmst}{SemiParametricModel}(object, class = c("rmst",
  "FlexTable")[2], digits = 3, pval_digits = digits, ...)
}
\arguments{
\item{object}{(SemiParametricModel) The object which was created when
fitting the Cox model}

\item{...}{additional parameters needed for specific instances of this
generic}

\item{class}{('rmst' or "FlexTable' (default)) type of output required}

\item{digits}{(numeric default=3) if outputting a FlexTable then the number of digits
to round the entries to}

\item{pval_digits}{(numeric default = same value as digits argument) decimal place rounding of p value}
}
\value{
(rmst object) contains list of RMST values, differences and call or
a FlexTable for output into a word document (depending on the class variable)
}
\description{
Method to calculate RMST on subset of data contained in a
SemiParametricModel object
}
