% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnDef.R
\docType{class}
\name{ColumnDef-class}
\alias{ColumnDef-class}
\title{Class representing a column of input data for the \code{SurvivalData} object}
\description{
Class representing a column of input data for the \code{SurvivalData} object
}
\section{Slots}{

\describe{
\item{\code{columnName}}{(character) The column name for which}

\item{\code{displayName}}{(character) The name to display when referring to this column in
output tables}

\item{\code{type}}{(One of "logical", "categorical", "numeric") The expected type of the column}

\item{\code{categories}}{(NULL or factor) if \code{type=="categorical"} this contains the 
set of allowed categories, ordered as required by the user}

\item{\code{unit}}{(character) The unit, if any, associated with the data in this column}
}}
\seealso{
\code{\link{ColumnDef}}
}

