% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnDef.R
\name{ColumnDef}
\alias{ColumnDef}
\title{Constructor for \code{ColumnDef} object}
\usage{
ColumnDef(columnName, type = NULL, categories = NULL, unit = "",
  displayName = NULL)
}
\arguments{
\item{columnName}{(character) name of column as it appears in the raw data}

\item{type}{(character) how to interpret values in the column,
e.g. "numeric" or "categorical" or "logical"}

\item{categories}{(NULL or factor) expected levels in a categorical column. If this
is entered as vector of character values, they will be coerced into factors}

\item{unit}{(character) unit of the column}

\item{displayName}{(character) name to use for column in outputs}
}
\value{
An \code{ColumnDef} object
}
\description{
Constructor for \code{ColumnDef} object
}
\examples{

#Define columns corresponding to arm
#coercing categories into factors
armDef <- ColumnDef(columnName = "grp",
                   type = "categorical",
                   categories = c("patchOnly", "combination"))

#Define columns corresponding to covariates
covariateDef <- list(
  ColumnDef(columnName = "age",
            type = "numeric",
            unit= "years"),
  
  ColumnDef(columnName = "race",
            type = "categorical",
            categories = factor(c("black", "hispanic", "other", "white"),
                                levels=c("black", "white", "hispanic", "other"))))


}
\seealso{
\code{\link{ColumnDef-class}}
}

