% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametric.R
\docType{methods}
\name{coxphLogRankTest}
\alias{coxphLogRankTest}
\alias{coxphLogRankTest,SemiParametricModel-method}
\title{Method to extract the logrank test from the Cox model
fit for the SemiParametricModel - this uses the strata and
covariates parameters used when creating the SemiParametricModel object}
\usage{
coxphLogRankTest(object)

\S4method{coxphLogRankTest}{SemiParametricModel}(object)
}
\arguments{
\item{object}{(SemiParametricModel) The object which was created when
fitting the Cox model}
}
\value{
A dataframe containing the log rank test results
}
\description{
Method to extract the logrank test from the Cox model
fit for the SemiParametricModel - this uses the strata and
covariates parameters used when creating the SemiParametricModel object
}

