% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageCurvePlotData.R
\docType{methods}
\name{createAvCurvePlotData}
\alias{createAvCurvePlotData}
\alias{createAvCurvePlotData,SurvivalModel-method}
\alias{createAvCurvePlotData,SurvivalModel-methods}
\title{Method to create the \code{AvCurvePlotData} object}
\usage{
createAvCurvePlotData(object, ...)

\S4method{createAvCurvePlotData}{SurvivalModel}(object, maxTime = NULL,
  Npoints = 201, Nsim = 500, models = NULL, seed = NULL)
}
\arguments{
\item{object}{\code{SurvivalModel} object}

\item{...}{Additional arguments for specific instances of this generic}

\item{maxTime}{(numeric) the time to extrapolate the curves to (default=NULL implies no extrapolation)}

\item{Npoints}{(numeric) the number of time points for which the survival curves are to be evaluated at}

\item{Nsim}{(numeric) the number of simulations used to generate the averaged curves}

\item{models}{(character vector) which models from \code{names(object@models)}) are to be used when 
calculating averaged survival curves - default NULL implies use all}

\item{seed}{(numeric, default NULL) if not NULL then set random seed}
}
\value{
A \code{AvCurvePlotData} object
}
\description{
Method to create the \code{AvCurvePlotData} object
}

