% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/icTable.R
\docType{methods}
\name{createIcTable}
\alias{createIcTable}
\alias{createIcTable,SurvivalModel-method}
\alias{createIcTable,SurvivalModel-methods}
\title{Method to create to create the information criteria table}
\usage{
createIcTable(object, ...)

\S4method{createIcTable}{SurvivalModel}(object, summaryFn = c("sum",
  "identity")[1], class = c("data.frame", "FlexTable")[2], digits = 2)
}
\arguments{
\item{object}{SurvivalModel object}

\item{...}{additional arguments for the generic}

\item{summaryFn}{("sum" or "identity") if "sum" the AIC and BIC are summed
over each arm producing a single value per model, it "identity" then the AIC
and BIC are output for each arm separately}

\item{class}{('data.frame' or (default) 'FlexTable') whether to output a data frame or FlexTable}

\item{digits}{(numeric default 2) The number of decimal places to round the AIC/BIC to}
}
\value{
A table contain the AIC and BIC for each model
}
\description{
Method to create to create the information criteria table
}

