% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTable.R
\docType{methods}
\name{createLifeTable}
\alias{createLifeTable}
\alias{createLifeTable,SurvivalModel-method}
\alias{createLifeTable,SurvivalModel-methods}
\title{Method to create life table}
\usage{
createLifeTable(object, ...)

\S4method{createLifeTable}{SurvivalModel}(object, times, modelToUse = NULL,
  Nsim = 500, class = c("data.frame", "FlexTable")[2], digits = 3,
  seed = NULL)
}
\arguments{
\item{object}{(SurvivalModel object) contains fitted models}

\item{...}{additional arguments for specific instances of this generic}

\item{times}{(numeric vector) times at which to evaluate life table}

\item{modelToUse}{(character) name of model to base life table upon. If
NULL, the model with lowest AIC is chosen}

\item{Nsim}{(numeric) number of simulations used to estimate the average
curve}

\item{class}{(data.frame or default flexTable) the class of the created table}

\item{digits}{(numeric default 3) the number of decimal places to round the
FlexTable output to}

\item{seed}{(numeric, default NULL) if not NULL then set random seed}
}
\value{
The lifetable
}
\description{
Method to create life table
}

