% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineAICtable.R
\docType{methods}
\name{createSplineAICTable}
\alias{createSplineAICTable}
\alias{createSplineAICTable,SurvivalModel-method}
\alias{createSplineAICTable,SurvivalModel-methods}
\title{Method to create the AIC table for the different spline models}
\usage{
createSplineAICTable(object, ...)

\S4method{createSplineAICTable}{SurvivalModel}(object, k = 2:5,
  scale = c("hazard", "odds", "normal"), class = c("data.frame",
  "FlexTable")[2], digits = 2)
}
\arguments{
\item{object}{\code{SurvivalModel} contains data to be fitted}

\item{...}{additional arguments for specific instances of this generic (}

\item{k}{(numeric vector) The number of knots to be used when fitting spline model
(if k=c(2,3) then models will be fit with both 2 and 3 knots)}

\item{scale}{(character vector) The scale arguement to be passed to flexsurvspline
(if scale=c("hazard", "odds") then models will be fitted with both hazard and odds)}

\item{class}{('data.frame' or 'FlexTable') the output data type}

\item{digits}{(numeric) Number of digits to round the flexTable output to}
}
\value{
A data frame or FlexTable object
}
\description{
Method to create the AIC table for the different spline models
}

