% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalModels.R
\docType{methods}
\name{fitModels}
\alias{fitModels}
\alias{fitModels,SurvivalData-method}
\alias{fitModels,SurvivalData-methods}
\title{Method to create the \code{SurvivalModel} object}
\usage{
fitModels(object, ...)

\S4method{fitModels}{SurvivalData}(object, armAsFactor = TRUE,
  covariates = character(0), endPoint, subgroup = as.character(NA),
  model = c("exponential", "weibull", "llogis", "lnorm", "gompertz"),
  modelOptions = NULL, preferredPackage = getDefaultPackage())
}
\arguments{
\item{object}{\code{SurvivalData} contains data to be fitted}

\item{...}{additional arguments for specific instances of this generic}

\item{armAsFactor}{(logical) TRUE if arm is to be included in the model,
FALSE if separate models are to be fitted for each arm}

\item{covariates}{(character vector) names of covariates to use}

\item{endPoint}{(character) name of the end-point to be fitted}

\item{subgroup}{(character string or NA) name of subgroup to fit}

\item{model}{(character vector) names of model distributions to fit}

\item{modelOptions}{(named list) extra options to be passed to the fitting functions.
For example modelOptions=list(spline=list(k=4)) will pass the argument k=4 to flexsurvSpline}

\item{preferredPackage}{(character) name of the preferred package in which
to look for the specified model}
}
\value{
A \code{SurvivalModel} object
}
\description{
Method to create the \code{SurvivalModel} object
}
