% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/semiParametric.R
\docType{methods}
\name{fitSemiParametric}
\alias{fitSemiParametric}
\alias{fitSemiParametric,SurvivalData-method}
\title{Method to fit semi-parametric model (Cox) and Kaplan-Meier
estimator to a SurvivalData object}
\usage{
fitSemiParametric(object, ...)

\S4method{fitSemiParametric}{SurvivalData}(object, endPoint,
  subgroup = as.character(NA), covariates = character(0),
  strata = character(0))
}
\arguments{
\item{object}{(SurvivalData) The object to fit the semi-parametric model
and calculate the KM estimator}

\item{...}{additional argument}

\item{endPoint}{(character) which endpoint the SemiParametricModel is associated with}

\item{subgroup}{(character) which subgroup the SemiParametricModel is associated with (or \code{NA}
if not associated with any)}

\item{covariates}{(character vector) names of covariates used when fitting the Cox model if any}

\item{strata}{(character vector) names of covariates to be stratified when fitting the Cox model}
}
\description{
Method to fit semi-parametric model (Cox) and Kaplan-Meier
estimator to a SurvivalData object
}

