% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTable.R
\name{getBestModel}
\alias{getBestModel}
\title{Output the model distribution with the lowest AIC}
\usage{
getBestModel(object)
}
\arguments{
\item{object}{(SurvivalModel)}
}
\value{
The name of the model with the lowest AIC
}
\description{
Output the model distribution with the lowest AIC
}
\details{
If armAsFactor is FALSE and separate models are
fitted then for each distribution calculate the sum of the
AIC and returns the distribution with the lowest sum of AIC
}
