% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCoeffs.R
\name{getDistributionDisplayNames}
\alias{getDistributionDisplayNames}
\title{Convert the distribution names from FlexSurv (e.g. `lnorm')
to display names (e.g. Lognormal)}
\usage{
getDistributionDisplayNames(modelNames)
}
\arguments{
\item{modelNames}{(character vector) A list of distribution names}
}
\value{
A vector of display names
}
\description{
Convert the distribution names from FlexSurv (e.g. `lnorm')
to display names (e.g. Lognormal)
}
\details{
If name is not  lnorm, llogis, gengamma then the function returns
the name unchanged except the first character capitalized.
}

