% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sibyl-package.R, R/survivalData.R,
%   R/survivalModels.R, R/averageCurvePlotData.R, R/semiParametric.R
\docType{methods}
\name{getEndpointUnits}
\alias{getEndpointUnits}
\alias{getEndpointUnits,SurvivalData-method}
\alias{getEndpointUnits,SurvivalData-methods}
\alias{getEndpointUnits,SurvivalModel-method}
\alias{getEndpointUnits,SurvivalModel-methods}
\alias{getEndpointUnits,AvCurvePlotData-method}
\alias{getEndpointUnits,AvCurvePlotData-methods}
\alias{getEndpointUnits,SemiParametricModel-method}
\alias{getEndpointUnits,SemiParametricModel-methods}
\title{Method to extract the endpoint units for a given object}
\usage{
getEndpointUnits(object)

\S4method{getEndpointUnits}{SurvivalData}(object)

\S4method{getEndpointUnits}{SurvivalModel}(object)

\S4method{getEndpointUnits}{AvCurvePlotData}(object)

\S4method{getEndpointUnits}{SemiParametricModel}(object)
}
\arguments{
\item{object}{The object whose endpoint units is required}
}
\value{
The units of the endpoints for the given object
}
\description{
Method to extract the endpoint units for a given object
}
