% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametric.R
\name{getMaxEndpointTime}
\alias{getMaxEndpointTime}
\alias{getMaxEndpointTime,SemiParametricModel-method}
\title{Method to calculate max endpoint time for semiparametric data}
\usage{
getMaxEndpointTime(object)
}
\description{
Method to calculate max endpoint time for semiparametric data
}
