% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFit.R
\name{getSplineKnotLocations}
\alias{getSplineKnotLocations}
\title{Given a SurvivalModel object, output a data frame
containing the knot locations (NOT on log-scale) of the given model name}
\usage{
getSplineKnotLocations(object, k, scale, class = c("data.frame",
  "FlexTable")[2], digits = 5)
}
\arguments{
\item{object}{(SurvivalModel) Survival model containing the
models fit}

\item{k}{(numeric) The number of knots of the spline model
whose knot locations are required}

\item{scale}{("hazard", "odds", or "normal") The scale argument
of the spline model whose knots are required}

\item{class}{("data.frame" or "FlexTable") whether to output the table
as a data.frame or FlexTable}

\item{digits}{(numeric) The number of digits to round the locations
when class="FlexTable"}
}
\description{
Given a SurvivalModel object, output a data frame
containing the knot locations (NOT on log-scale) of the given model name
}

