% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalData.R
\name{getZeroTimes}
\alias{getZeroTimes}
\title{Output details of subjects with zero times
for any endpoint}
\usage{
getZeroTimes(object)
}
\arguments{
\item{object}{(SurvivalData object)}
}
\value{
(character) containing details of subjects with zero times
or NULL if no {subject, endpoint} pairs with zero times
}
\description{
Output details of subjects with zero times
for any endpoint
}

