% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sibyl-package.R, R/survivalData.R,
%   R/survivalModels.R, R/averageCurvePlotData.R, R/semiParametric.R
\docType{methods}
\name{isSingleArm}
\alias{isSingleArm}
\alias{isSingleArm}
\alias{isSingleArm,SurvivalData-method}
\alias{isSingleArm}
\alias{isSingleArm,SurvivalModel-method}
\alias{isSingleArm}
\alias{isSingleArm,AvCurvePlotData-method}
\alias{isSingleArm}
\alias{isSingleArm,SemiParametricModel-method}
\title{Method to check whether an object contains data/models for a one arm clinical
trial}
\usage{
isSingleArm(object)

\S4method{isSingleArm}{SurvivalData}(object)

\S4method{isSingleArm}{SurvivalModel}(object)

\S4method{isSingleArm}{AvCurvePlotData}(object)

\S4method{isSingleArm}{SemiParametricModel}(object)
}
\arguments{
\item{object}{The object to check whether it contains the details for a
single arm trial}
}
\value{
logical, whether the object was a single arm trial
}
\description{
Method to check whether an object contains data/models for a one arm clinical
trial
}
