% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larvae}
\alias{larvae}
\title{larvae
This is a Phyloseq object containing example data from a 16S amplicon study
performed on mosquito larvae from Ethiopia.}
\format{
An object of class \code{phyloseq} of length 1.
}
\usage{
larvae
}
\description{
larvae
This is a Phyloseq object containing example data from a 16S amplicon study
performed on mosquito larvae from Ethiopia.
}
\keyword{datasets}
