% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalData.R
\name{maxObservedTime}
\alias{maxObservedTime}
\title{Calculate the maximum observed time 
for a given endPoint}
\usage{
maxObservedTime(object, endPointName)
}
\arguments{
\item{object}{(SurvivalData object)}

\item{endPointName}{(character) An endpoint associated with the
SurvivalData object}
}
\value{
(numeric) maximum observed time
}
\description{
Calculate the maximum observed time 
for a given endPoint
}

