% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalData.R
\name{minOfMaxObserved}
\alias{minOfMaxObserved}
\title{Calculate the maximum observed time for each arm
for a given endPoint and output the minimum}
\usage{
minOfMaxObserved(object, endPointName)
}
\arguments{
\item{object}{(SurvivalData object)}

\item{endPointName}{(character) An endpoint associated with the
SurvivalData object}
}
\value{
(numeric) minimum of the maximum observed time
for each arm
}
\description{
Calculate the maximum observed time for each arm
for a given endPoint and output the minimum
}

