% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{ord_and_mean}
\alias{ord_and_mean}
\title{Perform Ordination and Compute Consensus Coordinates}
\usage{
ord_and_mean(rarefied_matrix_list, repeats, cores = 4)
}
\arguments{
\item{rarefied_matrix_list}{A list of rarefied count tables.}

\item{repeats}{An integer. The number of rarefaction repeats.}

\item{cores}{An integer. The number of cores to use. Default = 4.}
}
\value{
A list containing:
\itemize{
\item \code{aligned_ordinations}: List of aligned ordinations.
\item \code{consensus_coordinates}: Consensus coordinates using Procrustes alignment.
}
}
\description{
Computes ordination (PCoA) using Bray-Curtis distance and aligns
results across multiple rarefied datasets.
This is the second step of the algorithm.
}
