% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageCurvePlotData.R, R/semiParametricOutput.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,AvCurvePlotData,missing-method}
\alias{plot,SemiParametricModel,missing-method}
\title{Plot methods for Sibyl package}
\usage{
\S4method{plot}{AvCurvePlotData,missing}(x, use.facet = TRUE,
  outputModel = NULL, xMax = NULL, useCI = FALSE)

\S4method{plot}{SemiParametricModel,missing}(x, type = c("KM", "CumHaz",
  "LoglogS", "LogoddS", "InvNormS")[1], use.facet = TRUE, logTime = NULL,
  ...)
}
\arguments{
\item{x}{(SemiParametricModel object) contains data to be plotted}

\item{use.facet}{(logical) should the treatment arms be split into different facets}

\item{outputModel}{(character) which model's survival curves should be plotted (default NULL
implies use all) if =character(0) then output only the KM curve}

\item{xMax}{(numeric or default NULL) the x-axis limit for the graph. If not included then
all data is displayed}

\item{useCI}{(logical) TRUE if including CI on graph, FALSE otherwise}

\item{type}{(character) the type of plot to be created; one of "KM",
"CumHaz", "LoglogS", "LogoddS" or "InvNormS"}

\item{logTime}{(logical) determines if x axis of plot is time or log(time)}

\item{...}{arguments to be passed to azplot.km when type = "KM"}
}
\description{
Plot methods for Sibyl package
}

