% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{plot_rep_raref}
\alias{plot_rep_raref}
\title{Generate a Plot for Repeated Rarefaction}
\usage{
plot_rep_raref(
  aligned_ordinations,
  consensus_coordinates,
  info,
  color,
  group,
  cloud,
  ellipse,
  title
)
}
\arguments{
\item{aligned_ordinations}{A list of aligned ordination matrices.}

\item{consensus_coordinates}{A matrix of consensus coordinates.}

\item{info}{A data frame containing sample metadata.}

\item{color}{A string. Column name in \code{info} for coloring.}

\item{group}{A string. Column name in \code{info} for grouping.}

\item{cloud}{A boolean. If \code{TRUE}, all repeated points are shown. Default = FALSE.}

\item{ellipse}{A boolean. If \code{TRUE}, confidence ellipses are drawn. Default = TRUE.}

\item{title}{A string. Title of the plot.}
}
\value{
A list containing:
\itemize{
\item \code{plot}: The generated ggplot2 plot.
\item \code{consensus_df}: A data frame of consensus coordinates.
\item \code{df_all}: A data frame of all ordination positions.
\item \code{updated_info}: Updated metadata.
}
}
\description{
Creates a PCA-based visualization of rarefied ordinations, highlighting consensus points.
}
