% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulation_test.R
\name{print.accumulation_test}
\alias{print.accumulation_test}
\title{Print Method for \code{accumulation_test} Object}
\usage{
\method{print}{accumulation_test}(x, ...)
}
\arguments{
\item{x}{An \code{accumulation_test} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for \code{accumulation_test} Object
}
