% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_threshold.R
\name{print.test_threshold}
\alias{print.test_threshold}
\title{Print Method for \code{test_threshold} Objects}
\usage{
\method{print}{test_threshold}(x, ...)
}
\arguments{
\item{x}{A \code{test_threshold} object.}

\item{...}{Additional arguments (not used).}
}
\description{
Prints the index plot from a \code{test_threshold} object.
}
