% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survivalModels.R
\docType{methods}
\name{removeModel}
\alias{removeModel}
\alias{removeModel,SurvivalModel-method}
\alias{removeModel,SurvivalModel-methods}
\title{Method to remove fitted models from a SurvivalData object}
\usage{
removeModel(object, ...)

\S4method{removeModel}{SurvivalModel}(object, modelName)
}
\arguments{
\item{object}{(SurvivalModel object) contains list of already fitted models}

\item{...}{additional arguments for specific instances of this generic}

\item{modelName}{(list of character strings) names of models to be removed. To remove
specific spline model should include the number of knots and scale. For example, to remove
the spline model with 5 knots and scale = hazard, should use \code{spline_5_hazard}}
}
\value{
a \code{SurvivalModel} object
}
\description{
Method to remove fitted models from a SurvivalData object
}

