% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{repeated_rarefaction}
\alias{repeated_rarefaction}
\title{Perform repeated rarefaction and produce an ordination plot based on PCA.}
\usage{
repeated_rarefaction(
  input,
  repeats = 50,
  threshold = 250,
  colorb = "sample_id",
  group = "sample_id",
  cloud = TRUE,
  ellipse = FALSE,
  cores = 4
)
}
\arguments{
\item{repeats}{A positive integer. Indicates the amount of repeats run.
A value = 1 means no repeats are used. Default = 50.}

\item{threshold}{A positive integer. The threshold value for the rarefaction. Default = 250}

\item{colorb}{A string. Name of the column with data to color the samples by.}

\item{group}{A string. Name of the column with data to group the samples by, this is also gonna
be the group the ellipse calculation is based on.}

\item{cloud}{Boolean. Aesthetic setting for the graph. Default is FALSE.
TRUE shows datapoints for all repeats.}

\item{ellipse}{Boolean. Aesthetic setting for the graph. Default is TRUE.}

\item{cores}{An int. The amount of cores used when running. The default = 4.}

\item{physeq}{A phyloseq object.}
}
\value{
A list with repeat count table, repeat info table, ordination object,
physeq object, dataframe with all repeat ordination positions, dataframe with median ordination positions, and ordination plot.
}
\description{
Perform repeated rarefaction and produce an ordination plot based on PCA.
}
\examples{
repeated_rarefaction(HLCYG_physeq_data, repeats = 100, threshold = 1000, colorb = "location", group = "location", ellipse = TRUE)
}
