% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenfeld.R
\docType{methods}
\name{schoenfeldResiduals}
\alias{schoenfeldResiduals}
\alias{schoenfeldResiduals,SemiParametricModel-method}
\title{Method to calculate the Schoenfeld residuals for a SemiParametricModel
object.}
\usage{
schoenfeldResiduals(object, ...)

\S4method{schoenfeldResiduals}{SemiParametricModel}(object, ...)
}
\arguments{
\item{object}{(SemiParametricModel) The object for which to}

\item{...}{additional arguments passed to cox.zph}
}
\value{
An object of class \code{cox.zph}
}
\description{
Method to calculate the Schoenfeld residuals for a SemiParametricModel
object.
}
\details{
Note there is no stratification used when calculating fitting 
the Cox model
}

