% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametric.R
\docType{class}
\name{SemiParametricModel-class}
\alias{SemiParametricModel-class}
\title{SemiParametric class definition}
\description{
SemiParametric class definition
}
\details{
Class storing the fitted Cox semi-parametric model and the
Kaplan Meier estimate (from survfit) for a given subgroup and endpoint
for a SurvivalData object.
}
\section{Slots}{

\describe{
\item{\code{cox}}{(coxph or NULL) The Cox model fit for the given subgroup and endpoint fitted using \code{coxph}
and no stratifiation or covariate variables. NULL if single arm trial}

\item{\code{coxWithStrata}}{(coxph or Null) The Cox model fit for the given subgroup and endpoint fitted 
using \code{coxph} and the selected stratifiation and covariate variable. 
If no strata included then this is the same as slot cox. NULL if single arm trial and no covariates/strata}

\item{\code{km}}{(survfit) The Kaplan-Meier estimator associated with the given subgroup and endpoint
generated by calling \code{survfit}. Note this estimator does NOT include the covariates
and strata included with the Cox model fit}

\item{\code{survData}}{(SurvivalData) object used when fitting the models. NOTE: if a subgroup
was selected when creating this object then only subjects in the given subgroup who have data for the
given endpoint will be included in this object}

\item{\code{subgroup}}{(character) which subgroup the SemiParametricModel is associated with (or \code{NA}
if not associated with any)}

\item{\code{endPoint}}{(character) which endpoint the SemiParametricModel is associated with}

\item{\code{covariates}}{(character vector) names of covariates used when fitting the Cox model}

\item{\code{strata}}{(character vector) names of covariates to be used to stratify the Cox model fit}

\item{\code{endPointDef}}{(list) defines name, time and censor column for the fitted
endpoint}
}}

\seealso{
\code{fitSemiParametric} for further information
}
