% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnDef.R, R/survivalData.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,ColumnDef-method}
\alias{show,SurvivalData-method}
\title{Show methods for Sibyl objects}
\usage{
\S4method{show}{ColumnDef}(object)

\S4method{show}{SurvivalData}(object)
}
\arguments{
\item{object}{(SurvivalData object)}
}
\description{
Show methods for Sibyl objects
}

