\name{sibyl-package}
\alias{sibyl-package}
\alias{sibyl}
\docType{package}
\title{
\packageTitle{sibyl}
}
\description{
\packageDescription{sibyl}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{sibyl}
\packageIndices{sibyl}
You define a system, namely vars and compos.(def.var, def.compo)
and define what vars you use in te time series data and the start date-time.(wather$new)
You run sibyl.(sibyl)
then get a csv file contains data.

}
\author{
\packageAuthor{sibyl}

Maintainer: \packageMaintainer{sibyl}
}
\references{

\keyword{ package }
\seealso{
}
\examples{

}
