% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Sibyl-package.R
\docType{package}
\name{Sibyl}
\alias{Sibyl}
\alias{Sibyl-package}
\title{Sibyl: The package fits parametric models to time-to-event data and
extrapolates these curves to larger surivival times.}
\description{
Sibyl: The package fits parametric models to time-to-event data and
extrapolates these curves to larger surivival times.
}

