% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sibyl-package.R
\docType{data}
\name{sibylData}
\alias{sibylData}
\title{Example time to event data set for subjects trying to stop smoking}
\format{A data frame with the following columns:
\describe{
  \item{ID}{the unique subject id}
  \item{age}{age of the subject, in years}
  \item{race}{'white', 'black', 'hispanic' or 'other'}
  \item{grp}{'combination' or 'patchOnly'; the name of the treatment arm to which the subject was assigned}
  \item{sub.isMale}{ndicator for membership of the subgroup 'isMale'}
  \item{sub.isHeavySmoker}{indicator for membership of the subgroup 'isHeavySmoker'}
  \item{ttr}{time to relapse, in days}
  \item{ttr.cens}{indicator for relapse}
  \item{end.2}{additional endpoint}
  \item{cens.2}{indicator for additional endpoint}
}}
\usage{
sibylData
}
\description{
Example time to event data set for subjects trying to stop smoking
}
\details{
This data set is loosely based on an example
data frame from the asaur package
}
\keyword{datasets}

