% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificEndpointValidation.R
\name{specificEndpointRules}
\alias{specificEndpointRules}
\title{Produce warning messages if SurvivalData object endpoints
do not satisfy a set of rules}
\usage{
specificEndpointRules(object, rulesToCheck = NULL)
}
\arguments{
\item{object}{(SurvivalData object)}

\item{rulesToCheck}{(numeric vector) - which rules should be checked
NULL if check all rules}
}
\value{
A string of warning messages or NULL if no warnings
}
\description{
Produce warning messages if SurvivalData object endpoints
do not satisfy a set of rules
}
\details{
Rules:
1) If endpoint PFS2 and PFS exists then for all subjects time to PFS must be < time
to PFS2 
2) If endpoint OS exists then for all subjects who have an event for OS, their event
time for OS must be >= their time for all other endpoints
}
