% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCoeffs.R
\docType{methods}
\name{summariseCoeffs}
\alias{summariseCoeffs}
\alias{summariseCoeffs,SurvivalModel-method}
\title{Method to create summary tables of coefficients for each model}
\usage{
summariseCoeffs(object, ...)

\S4method{summariseCoeffs}{SurvivalModel}(object, class = c("matrix",
  "FlexTable")[2], digits = 5)
}
\arguments{
\item{object}{(SurvivalModel object) contains fitted models to be summarised}

\item{...}{additional arguments to summariseCoeffs}

\item{class}{('matrix' or 'FlexTable' (default)) output format
for the summary tables}

\item{digits}{(numeric) The number of significant digits to round the FlexTable output to.
This option is ignored if class is not 'FlexTable'}
}
\description{
Method to create summary tables of coefficients for each model
}

