% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametricOutput.R, R/survivalDataOutput.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,SemiParametricModel-method}
\alias{summary,SurvivalData-method}
\title{Summary methods for Sibyl objects}
\usage{
\S4method{summary}{SemiParametricModel}(object, class = c("data.frame",
  "FlexTable")[2], digits = 1)

\S4method{summary}{SurvivalData}(object, type = c("subgroups", "endPoints",
  "covariates")[1], digits = 1, htmlEncoding = FALSE,
  meanOrMedian = c("median", "mean")[1])
}
\arguments{
\item{object}{(SurvivalData object) data to be summarised}

\item{class}{('data.frame' or "FlexTable' (default)) type of output for summary table}

\item{digits}{(numeric, default 1) number of decimal places for the output}

\item{type}{(one of "subgroups", "endpoints", "covariates") which output table
should be generated for the \code{SurvivalData} object}

\item{htmlEncoding}{(logical) if TRUE use &plusmn instead of unicode plus minus sign}

\item{meanOrMedian}{("median" or "mean") output summary function for the numeric covariates
summary table}
}
\description{
Summary methods for Sibyl objects
}

