% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalData.R
\docType{class}
\name{SurvivalData-class}
\alias{SurvivalData-class}
\title{Class representing the data to be used by Sibyl for model fitting
and extrapolating.}
\description{
Class representing the data to be used by Sibyl for model fitting
and extrapolating.
}
\section{Slots}{

\describe{
\item{\code{subject.data}}{(data frame), one row per subject
with unique subject ID, their covariate values, treatment arm,
time to event and censor indicators for multiple endpoints and
indicators as to which subgroups the subject belongs.
Colums names are described by the data in other slots (although the treatment arm
column is named "arm").}

\item{\code{armDef}}{(\code{ColumnDef}) object describing the treatment arm column and its categories
(the control group should be first) in the subject.data data frame}

\item{\code{endPoints}}{(list of lists) each endpoint is the name of a two element list, "timeCol" - the time
to event for the given endpoint and "censorCol" whether the subject was censored (\code{TRUE}) or not (\code{FALSE}).}

\item{\code{endPointUnit}}{(days, weeks, months, years) The units for the endpoints}

\item{\code{subgroupDef}}{(list of \code{ColumnDef}) definitions of the columns in the subject.data data frame
describing subgroups}

\item{\code{covDef}}{(list of \code{ColumnDef}) definitions of columns in the
subject.data data frame describing covariates}
}}
\seealso{
\code{\link{SurvivalData}}
}

