% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalData.R
\name{SurvivalData}
\alias{SurvivalData}
\title{Constructor for \code{SurvivalData} object}
\usage{
SurvivalData(data, armDef, subjectCol, covDef = NULL, subgroupDef = NULL,
  endPointNames, censorCol, timeCol, endPointUnit = c("days", "months",
  "years")[2])
}
\arguments{
\item{data}{(data frame) raw data to be formatted into SurvivalData object}

\item{armDef}{(columnDef object) definition of the data column that
specifies the trial arm each subject was on}

\item{subjectCol}{(string) name of the data column containing the unique subject
identifier}

\item{covDef}{(list of columnDef objects, default=NULL) definition of each
required covariate, including name, type, unit, etc.}

\item{subgroupDef}{(list of columnDef objects, default=NULL) definition of
data columns that indicate membership of a subgroup.}

\item{endPointNames}{(vector of strings) names of the endpoints}

\item{censorCol}{(vector of strings) names of the data columns that specify
which subjects are censored for a particular end point. The N-th
element should correspond to the N-th end point.}

\item{timeCol}{(vector of strings) names of the data columns that specify
the times at which subjects reached a particular end point. The N-th
element should correspond to the N-th end point.}

\item{endPointUnit}{("days", "months" or "years" - default "months") The unit of time
for the endPoint time columns}
}
\value{
A \code{SurvivalData} object
}
\description{
Constructor for \code{SurvivalData} object
}
\seealso{
\code{\link{SurvivalData-class}}
}

