% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalModels.R
\docType{class}
\name{SurvivalModel-class}
\alias{SurvivalModel-class}
\title{Class storing the fitted survival models for a specific subgroup, end point
and armAsFactor}
\description{
Class storing the fitted survival models for a specific subgroup, end point
and armAsFactor
}
\section{Slots}{

\describe{
\item{\code{models}}{(list) names of model distributions fitted}

\item{\code{covariates}}{(character vector) names of covariates used}

\item{\code{survData}}{(SurvivalData object) contains data for the selected subgroup
only}

\item{\code{armAsFactor}}{(logical) TRUE if arm is included in the model, FALSE if
separate models are fit for each arm}

\item{\code{survFormula}}{(formula) formula used to fit the model}

\item{\code{subgroup}}{(character string or NA) name of subgroup fitted}

\item{\code{endPointDef}}{(list) defines time and censor column for the fitted
endpoint}

\item{\code{endPoint}}{(character) name of the end-point to be fitted
see vignette and \code{SurvivalModel} for further iunformation}
}}

