% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCoeffs.R
\docType{methods}
\name{vcov}
\alias{vcov}
\alias{vcov,SurvivalModel-method}
\alias{vcov,SurvivalModel-methods}
\title{Method to return covariance matrix of estimated model parameters}
\usage{
vcov(object, ...)

\S4method{vcov}{SurvivalModel}(object, class = c("matrix", "FlexTable")[2],
  digits = 5)
}
\arguments{
\item{object}{(SurvivalModel object) contains fitted models with estimated
parameters}

\item{...}{additional arguments to vcov}

\item{class}{('matrix' or 'FlexTable' (default)) output format
for the summary tables}

\item{digits}{(numeric) The number of significant digits to round the FlexTable output to.
This option is ignored if class is not 'FlexTable'}
}
\description{
Method to return covariance matrix of estimated model parameters
}

