% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumulation_test.R
\name{accumulation_test}
\alias{accumulation_test}
\title{Accumulation Curve Analysis}
\usage{
accumulation_test(input, step = 5)
}
\arguments{
\item{input}{A \code{phyloseq} object.}

\item{step}{A numeric value. The step size for drawing the accumulation curve.
Default = 5.}
}
\value{
A list containing:
\itemize{
\item \code{accumulation_plot}: A ggplot2 object visualizing the accumulation curves.
\item \code{threshold_density}: A ggplot2 density plot of the 75\% ACE threshold.
}
}
\description{
This function generates accumulation (rarefaction) curves for each sample in
a given \code{phyloseq} object, fits a general accumulation model using the Abundance
Coverage Estimator (ACE) as an asymptote, and identifies the sequencing
depth at which 75\% of the ACE value is reached. It also produces a density
plot showing the distribution of these 75\% completion thresholds.
}
