% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adults}
\alias{adults}
\title{adults}
\format{
A \code{phyloseq} object with 34 samples and 5221 taxa.

A sample_data object is included, with the following columns:
\describe{
\item{sample_id}{A unique identifier for each sample.}
\item{location}{The location where the sample was collected.}
}
}
\source{
\url{https://pmc.ncbi.nlm.nih.gov/articles/PMC4785398/}
}
\usage{
adults
}
\description{
This is a \code{phyloseq} containing example data from
a 16S study performed on adult mosquitoes from Burkina Faso.
}
\keyword{datasets}
