% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pairwise_dist_per_sample.R
\name{avg_pairwise_dist_per_sample}
\alias{avg_pairwise_dist_per_sample}
\title{Calculate Average Pairwise Distance for Sample Clusters Across Rarefaction Thresholds}
\usage{
avg_pairwise_dist_per_sample(
  input,
  repeats = 50,
  t_min = 50,
  t_max = 250,
  t_step = 5,
  cores = 4
)
}
\arguments{
\item{input}{A \code{phyloseq} object.}

\item{repeats}{An integer or a vector of integers.
The number of times to repeat rarefaction. A value of 1 means no repeats.
If using a vector, different rarefaction thresholds will be tested sequentially.
Default = 50.}

\item{t_min}{An integer. The minimum value for the threshold testing range.
Default = 50.}

\item{t_max}{An integer. The maximum value for the threshold testing range.
Default = 250.}

\item{t_step}{A numeric value. The step size for the threshold testing range.
A value between 0 and 1 will cause the same threshold to be tested multiple times.
Default = 5.}

\item{cores}{An integer. The number of cores for parallel processing. Default = 4.}
}
\value{
A list containing:
\itemize{
\item \code{plot}: A ggplot2 object showing average pairwise distances across thresholds.
\item \code{distances}: A data frame with calculated distances per sample.
}
}
\description{
This function performs repeated rarefaction on a \code{phyloseq} object, computes
ordination, and calculates the average pairwise distance for each sample
across different rarefaction thresholds.
}
