% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pairwise_dist_plot.R
\name{avg_pairwise_dist_plot}
\alias{avg_pairwise_dist_plot}
\title{Plot average pairwise distance across thresholds}
\usage{
avg_pairwise_dist_plot(avg_distances)
}
\arguments{
\item{avg_distances}{A dataframe of average distances obtained from \code{test_threshold}.}
}
\value{
A \code{ggplot} object showing average pairwise distances across thresholds.
}
\description{
This function it's a plotting function for \code{avg_distances} generated from
\link{test_threshold}.
}
\examples{
library(Sibyl)
# Creating a smaller subset of the data
adults_sub <- phyloseq::subset_samples(adults, location=="VK3")
result <- test_threshold(adults_sub, 
                         repeats = 10, 
                         t_min = 100, 
                         t_max = 500, 
                         t_step = 50, 
                         group = "location",
                         verbose = FALSE)
avg_pairwise_dist_plot(result$avg_distances$repeat_number_10)
}
