% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametric.R
\docType{methods}
\name{coxSnellRes}
\alias{coxSnellRes}
\alias{coxSnellRes}
\alias{coxSnellRes,SemiParametricModel-method}
\title{Method to calculate Cox-Snell residual}
\usage{
coxSnellRes(object)

\S4method{coxSnellRes}{SemiParametricModel}(object)
}
\arguments{
\item{object}{(SemiParametricModel) The object which was created when
fitting the Cox model}

\item{...}{additional parameters needed for specific instances of this
generic}
}
\value{
(survfit object) contains call and residuals
}
\description{
Method to calculate Cox-Snell residual
}
\details{
This is calculated using surv(t, e) ~ arm, with no
covariates or strata. Data is the subset of data contained in the
SemiParametricModel object used and the "ties" argument used by coxph is the value used 
when creating the SemiParametric object
}
