% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\docType{data}
\name{HLCYG_physeq_data}
\alias{HLCYG_physeq_data}
\title{HLCYG_physeq_data}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{location}{sample location}
}
\usage{
HLCYG_physeq_data
}
\description{
This is a dataset containing stuff.
}
\keyword{datasets}
