% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\docType{data}
\name{HLCYG_physeq_data}
\alias{HLCYG_physeq_data}
\title{HLCYG_physeq_data
This is a dataset containing stuff.}
\format{
An object of class \code{phyloseq} of length 1.
}
\usage{
HLCYG_physeq_data
}
\description{
HLCYG_physeq_data
This is a dataset containing stuff.
}
\keyword{datasets}
