% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{larvae}
\alias{larvae}
\title{larvae}
\format{
A \code{phyloseq} object with 54 samples and 722 taxa.

A sample_data object is included, with the following columns:
\describe{
\item{sample_id}{A unique identifier for each sample.}
\item{breeding_site_type}{Environmental classification of sites where
mosquitoes were collected}
}
}
\source{
\url{https://academic.oup.com/femsec/article/101/1/fiae161/7928135}
}
\usage{
larvae
}
\description{
This is a \code{phyloseq} containing example data from
a 16S study performed on adult mosquitoes larvae from Ethiopia.
}
\keyword{datasets}
