% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_rarefaction.R
\name{ord_and_mean}
\alias{ord_and_mean}
\title{Calculates ordination on the repeated count input and as well as a median result for each original datapoint.}
\usage{
ord_and_mean(rarefied_matrix_list, repeats, cores = 4)
}
\arguments{
\item{repeats}{An integer. The amount of repeats that has been performed on the data.}

\item{rarified_matrix_list}{A matrix. Contains the repeated count data.}
}
\value{
Returns a list containg an ordination object, a phyloseq object with
the repeated count data, a dataframe with all positions from the ordination
calculaction, and a datafram with just the median position from the calculation.
}
\description{
Calculates ordination on the repeated count input and as well as a median result for each original datapoint.
}
