% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_raref.R
\name{rep_raref}
\alias{rep_raref}
\title{Perform Repeated Rarefaction}
\usage{
rep_raref(count, threshold, repeats)
}
\arguments{
\item{count}{A matrix. OTU count table.}

\item{threshold}{An integer. The threshold for rarefaction.}

\item{repeats}{An integer. Number of repeats.}
}
\value{
A list containing:
\itemize{
\item \code{rarefied_matrix_list}: A list of rarefied matrices.
}
}
\description{
This function performs repeated rarefaction on a count table.
This is the first step of the algorithm.
}
