% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiParametricOutput.R, R/survivalDataOutput.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,SemiParametricModel-method}
\alias{summary}
\alias{summary,SurvivalData-method}
\title{Summary methods for Sibyl objects}
\usage{
\S4method{summary}{SemiParametricModel}(object, type = c("medianTTE",
  "KM")[1], class = c("data.frame", "FlexTable")[2], digits = if (type ==
  "medianTTE") 2 else 5)

\S4method{summary}{SurvivalData}(object, type = c("subgroups", "endPoints",
  "covariates", "covarMaturity")[1], digits = 1, htmlEncoding = FALSE,
  meanOrMedian = c("median", "mean")[1])
}
\arguments{
\item{object}{(SurvivalData or SemiParametricModel object) data to be summarised}

\item{type}{(one of "subgroups", "endpoints", "covariates", "covarMaturity") which output table
should be generated for the \code{SurvivalData} object OR (one of "medianTTE" or "KM") if 
object is of type SemiParametricModel}

\item{class}{('data.frame' or "FlexTable' (default)) type of output for summary table}

\item{digits}{(numeric, default 1) number of decimal places for the output}

\item{htmlEncoding}{(logical) if TRUE use &plusmn instead of unicode plus minus sign}

\item{meanOrMedian}{("median" or "mean") output summary function for the numeric covariates
summary table}
}
\description{
Summary methods for Sibyl objects
}
