% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_testing.R
\name{test_threshold}
\alias{test_threshold}
\title{This function is a wrapper designed around "repeated_rarefaction". It allows.}
\usage{
test_threshold(
  input,
  repeats = 10,
  t_min = 50,
  t_max = 250,
  t_step = 1,
  group = "sample_id",
  cores = 4
)
}
\arguments{
\item{input}{A phyloseq object or a table of OTU/ASV counts.}

\item{repeats}{A positive integer. Indicates the amount of repeats run. A value = 1 means no repeats are used.}

\item{t_min}{An integer. Minimum value for the threshold testing range.}

\item{t_max}{An integer. Maximum value for the threshold testing range.}

\item{t_step}{An float. Step value for the threshold testing range. A value between 0 and 1 will cause the same threshold to be tested more than once.}

\item{group}{A string. Name of the column with data to group the points by.}

\item{cores}{An int. The amount of cores used when running. The default = 4.}

\item{sample_id}{A string. Name of the column with sample IDs.}
}
\value{
A graph with rarefaction threshold as the x axis, and the y axis is the calculated Calinski-Harabasz F-statistic. Each dot indicates the clustering score for the repeated ordination on a single threshold value. The higher score the better.
}
\description{
This function is a wrapper designed around "repeated_rarefaction". It allows.
}
\examples{
test_threshold(HLCYG_physeq_data, repeats = 10, t_min = 200, t_max =1500, t_step = 5, group="location")
}
