% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_testing.R
\name{test_threshold}
\alias{test_threshold}
\title{Test Different Rarefaction Thresholds Using Repeated Rarefaction}
\usage{
test_threshold(
  input,
  repeats = 10,
  t_min = 50,
  t_max = 250,
  t_step = 1,
  group = "sample_id",
  cores = 4
)
}
\arguments{
\item{input}{A \code{phyloseq} object or a table of OTU/ASV counts.}

\item{repeats}{An integer. The number of rarefaction repeats. A value of 1 means no repeats. Default = 10.}

\item{t_min}{An integer. The minimum value for the threshold testing range.}

\item{t_max}{An integer. The maximum value for the threshold testing range.}

\item{t_step}{A numeric value. The step size for the threshold testing range. A value between 0 and 1 will cause the same threshold to be tested multiple times.}

\item{group}{A string. Name of the column to group points by.}

\item{cores}{An integer. Number of cores for parallel processing. Default = 4.}
}
\value{
A list containing:
\itemize{
\item \code{index_plot}: A ggplot2 object showing rarefaction threshold vs. clustering performance.
\item \code{ordination_plots}: A list of ordination plots for different threshold values.
}
}
\description{
This function is a wrapper around \code{repeated_rarefaction} that evaluates different rarefaction thresholds.
It runs repeated rarefactions on a phyloseq object and calculates clustering performance using the Calinski-Harabasz index.
}
